src/requirements.txt
# Python 11 for rotating addresses
aiohttp==3.8.4
beautifulsoup4==4.12.2
python-dotenv==1.0.0
python-slugify==8.0.1
stem==1.8.2
